Move Name::Observations test to separate file
